pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jobpipeline"
version = "0.1.0"
description = "Local-first desktop job search pipeline for Windows"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PySide6>=6.7",
  "httpx>=0.27",
  "beautifulsoup4>=4.12",
  "openpyxl>=3.1",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "black>=24.0",
  "ruff>=0.5",
]

[project.scripts]
jobpipeline-cli = "jobpipeline.core.cli:main"
jobpipeline-ui = "jobpipeline.app.main:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["jobpipeline*"]
exclude = ["data*", "app_data*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
